refactor: remove final dto multiple streams and pressures train
