+++ src/libecalc/domain/process/compressor/dto/__init__.py
@@ -1,5 +1,2 @@
 from .sampled import CompressorSampled
 from .stage import CompressorStage, InterstagePressureControl
-from .train import (
-    VariableSpeedCompressorTrainMultipleStreamsAndPressures,
-)
